refactor(ms2/sysadmin): sepparate function to get space representation
